Style elements and add toggle user group function
The viewer has been styled to be in line with the Twitch color pallete.
Additionally, functionality was added to allow the user to click
whether they want to see all users, or online online or offline users.
Clicking each respective button calls a Javascript function to hide or
show the appropriate users as well as set the current menu item
with a different color to differentiate.

TODO:

* Implement a better way to populate displayUsers function. Currently
there is a timeout delay on the populateObj function due to the async
API calls. However, if the API calls take longer than the timeout, no
data will be displayed.