Create Javascript promises for API calls
The previous attempt at using promises was corrected and re-written
from the ground up. It is now fully functioning. Each promise, when
resolved increases the counter in the then function. When the counter
reaches the total number of users it displays all of the data. This
allowed the removal of an arbitray setTimeout function.