Change promise to fulfill for each user
The displayUsers function no longer waits for all user properties to
be populated. Instead, the promise resolves and passes along the data
for the particular user. The displayUsers function now processes the
data for one user and appends the HTML for the one user, rather than
running another for loop. This should allow better efficiency and
ability to scale up without unnecessary delays.