Add interval update functionality
The Twitch viewer will now poll the API every 5 minutes and update
the object and visible data with any changes to each users online
status and game. In order to reduce code and unnecessary API calls,
users who do not exist or whose account is unavailable are excluded
from this API call.

Additionally, some of the code for the initial object creation was
cleaned up. All users now have a game property created. If they are not
streaming it is set to "Offline." Users who do not exist or are
unavailable later overwrite this value. By manually coding in this
property, it allowed the removal of a function within the initial
display data function.

TODO:

* Add a refresh icon to the menubar that will allow the user to
manually call the update functions.