Add user manual refresh button and scripting
The user can now manually click on the refresh icon to run the
requisite functions to update the online and game values of each user.
The page will also query for the information each 5 minutes.

Additionally, some styling for smaller screen sizes was added. This
included reducing the logo and text size for better information
density.