Add user type verification
Functionality was added to remove all users from the initial array
which are not of a string type. It will not remove users whose account
does not exist or is not available. That is still displayed in the
game section of the user display.

Additional comments were added to the top of the file for ease of use
and implementation.