Add initial functionality of local storage
The app now utilizes local storage to maintain the list of users. This
commit introduces a function which checks for the users key in local
storage. If it does not exist, it creates the user key and prepopulates
it with some users, and then fetches and displays the user data. If
there is local storage, it grabs the data from the storage, saves it to
a variable and then fetches and displays the user data.

Additionally, the verifyUsersType and populateObj functions have been
made more granular so that they only accept a single user. This will
allow the functions to be reused when adding a single user. The for
loops for initial display were moved to the initializeStorage function.

Initial functionality to add a user has been added, but still needs to
be fully built out.