Add functionality to add multiple users at once
The addUser function now will test if the value it is passed contains
commas or spaces, which are not allowed in Twitch usernames. If it
matches either of those, it will split on those characters, and parse
all of the items in the array as new users to display and store.