Add interface and functionality to remove user
An "x" is created in the top right corner of each user card upon
creation with funcitonality to fade in and out based on the hover state
of the respective user. If clicked on, the removeUser function will run
and remove the user from the display, as well as update the
local storage.