Add collapsible user input field to add new users
This adds the necessary UI hooks to add new users to the app and local
storage. The end user can now add their own users to track in the
Twitch viewer. Users can be added one at a time or multiple via comma
or space delimiters.

In order to have the field flow correctly, Bootstrap CSS was added to
the project. Several later CSS declarations had to be declared importan
to maintain their values and not be overriden by Bootstrap.