Remove deprecated function and update comments
The verifyUsersType function was removed as it is no longer needed.
Since all user additions with the included input method will be strings
by default, there is no need to verify that they are strings.

Additionally, the comments at the top of the file were updated to
reflect the use of localStorage for the project now. Some typos in
other comments throughout the file were also corrected.